requirements.txt
numpy
dlib
open-cv2
face-recognition
cmake